Added map elements to  timetable
