+++ src/components/navigation/components/index.js
@@ -1,3 +1,3 @@
 export { SideBar } from "./sideBar";
 export { InfoModal } from "./infoModal";
-export { Timetable } from "./timeTable";
+
