+++ src/components/navigation/components/sideBar/components/index.js
@@ -1,2 +1,3 @@
 export { SelectCities } from "./citiesSelector";
 export { SelectCity } from "./citySelector";
+export { Timetable } from "./timeTable";
