+++ src/components/navigation/components/sideBar/components/timeTable.css
@@ -7,12 +7,19 @@
   border-collapse: collapse;
 }
 h2.departure_Timetable {
-  padding: 2vmin;
-  width: 75%;
+  padding-bottom: 2vmin;
+  padding-left:1vmin;
+  padding-top: 1vmin;
+  padding-right: 2vmin;
+  width: 80%;
 }
 .trainNumber {
   color: #000082;
   background-color: #fff;
+}
+.departure_Timetable table{
+  max-height: 80%;
+  overflow-y: scroll;
 }
 .departure_Timetable td,
 .departure_Timetable th {
@@ -30,7 +37,13 @@ h2.departure_Timetable {
   background-color: #5454af;
 }
 
-.departure_Timetable .selected {
-  background-color: #b32428 !important;
+table.departure_Timetable tbody tr.selected td {
+  background-color: #b32428;
   color: white;
+}
+
+@media (max-width: 1000px) {
+  .departure_Timetable .statusColumn {
+    display: none; /* Hide status column on smaller screens */
+  }
 }
