+++ src/components/navigation/components/sideBar/sideBar.jsx
@@ -1,6 +1,5 @@
 import PropTypes from "prop-types";
 import "./sideBar.css";
-import { Timetable } from "../timeTable";
 
 import { SideContent } from "./sideContent";
 import { Radio } from "antd";
@@ -31,7 +30,6 @@ export const SideBar = ({ isMenuOpen }) => {
         </div>
         <SideContent />
       </div>
-      <Timetable stationName="Frankfurt(Main)Hbf" />
     </div>
   );
 };
