+++ src/components/navigation/components/sideBar/sideContent.jsx
@@ -1,7 +1,6 @@
 import { useSelector } from "react-redux";
 import { getMode } from "../../../../app/app-selectors";
-import { SelectCities, SelectCity } from "./components";
-import { Timetable } from "../timeTable";
+import { SelectCities, SelectCity, Timetable } from "./components";
 
 export const SideContent = () => {
   const mode = useSelector(getMode);
@@ -9,7 +8,7 @@ export const SideContent = () => {
   if (mode === "averageBetween") {
     return <SelectCities />;
   } else if (mode === "timeTable") {
-    return <Timetable stationName="Frankfurt(Main)Hbf" />;
+    return <Timetable />;
   } else {
     return <SelectCity />;
   }
